Refactor code into UserTrackingScrollContainer
Always highlight selected section when user clicked on it

Fix floating-point imprecision

Issue in comment seems to be fixed
